Add hover & focus state on pricing cards
