+++ src/stories/PricingCard.stories.js
@@ -1,19 +1,19 @@
 import React from 'react'
 import PricingCard from 'components/PricingCard'
-import getPricingPageData from '../../data/pricing'
+import Typography from 'components/Typography'
 
 export default {
   title: 'Components/PricingCard',
   component: PricingCard,
   parameters: {
     layout: 'padded',
-    backgrounds: { default: 'white' },
   },
   decorators: [
     Story => (
       <div
         style={{
           width: 250,
+          color: 'white',
         }}
       >
         <Story />
@@ -22,10 +22,15 @@ export default {
   ],
 }
 
-export const Default = () => {
-  const {
-    pricing: { payingPlans },
-  } = getPricingPageData()
+const Template = args => <PricingCard {...args} />
 
-  return <PricingCard plan={payingPlans[0]} />
+export const Default = Template.bind({})
+Default.args = {
+  children: <Typography>Hover me</Typography>,
+}
+
+export const OtherColor = Template.bind({})
+OtherColor.args = {
+  children: <Typography>Hover me</Typography>,
+  color: 'colors.hotPink',
 }
